Most of size gone and most of constructor use replaces.

+++ std-bits/table/src/main/java/org/enso/table/parsing/DateParser.java
@@ -13,7 +13,7 @@ public class DateParser extends BaseTimeParser {
   }
 
   @Override
-  protected Builder makeBuilderWithCapacity(int capacity, ProblemAggregator problemAggregator) {
+  protected Builder makeBuilderWithCapacity(long capacity, ProblemAggregator problemAggregator) {
     return Builder.getForType(DateType.INSTANCE, capacity, problemAggregator);
   }
 }
